requirements.txt.txt
streamlit
numpy
pandas
python-dotenv
openai
gtts
SpeechRecognition
pydub
